Add variant conversion field attribute
Adds the #[variant] attribute which could be used to customize the behavior
of `FromVariant` and `ToVariant` derive macros. This is useful for cases where
custom behavior is desired only for a few fields, for example because one of
the fields does not implement the respective traits.

Initially supported arguments are:

- from_variant_with
- to_variant_with
- with
