+++ test/src/lib.rs
@@ -273,6 +273,8 @@ fn test_derive_to_variant() -> bool {
         foo: T::A,
         bar: T,
         baz: ToVarEnum<T::B>,
+        #[variant(with = "variant_with")]
+        ptr: *mut (),
     }
 
     #[derive(Clone, Eq, PartialEq, Debug, ToVariant, FromVariant)]
@@ -292,21 +294,39 @@ fn test_derive_to_variant() -> bool {
         type B = bool;
     }
 
+    mod variant_with {
+        use gdnative::{FromVariantError, GodotString, ToVariant, Variant};
+
+        pub fn to_variant(_ptr: &*mut ()) -> Variant {
+            GodotString::from("*mut ()").to_variant()
+        }
+
+        pub fn from_variant(_variant: &Variant) -> Result<*mut (), FromVariantError> {
+            Ok(std::ptr::null_mut())
+        }
+    }
+
     let ok = std::panic::catch_unwind(|| {
         let data = ToVar::<f64> {
             foo: 42,
             bar: 54.0,
             baz: ToVarEnum::Foo(true),
+            ptr: std::ptr::null_mut(),
         };
         let variant = data.to_variant();
         let dictionary = variant.try_to_dictionary().expect("should be dictionary");
         assert_eq!(Some(42), dictionary.get(&"foo".into()).try_to_i64());
         assert_eq!(Some(54.0), dictionary.get(&"bar".into()).try_to_f64());
+        assert_eq!(
+            Some("*mut ()".into()),
+            dictionary.get(&"ptr".into()).try_to_string()
+        );
         let enum_dict = dictionary
             .get(&"baz".into())
             .try_to_dictionary()
             .expect("should be dictionary");
         assert_eq!(Some(true), enum_dict.get(&"Foo".into()).try_to_bool());
+
         assert_eq!(
             Ok(&data.baz),
             ToVarEnum::from_variant(&enum_dict.to_variant()).as_ref()
